pyproject.toml
[tool.poetry]
name = "economic_observer"
version = "0.0.1"
description = ""
authors = ["DoubleChuang <ethan9141@gmail.com>"]
readme = "README.md"
packages = [{include = "economic_light_observer"}]

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"
requests-html = "^0.10.0"
lxml-html-clean = "^0.1.1"
nest-asyncio = "^1.6.0"
pyppeteer = { git = "https://github.com/DoubleChuang/pyppeteer.git", branch = "feat/suport-arm-and-arm64" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
